feat: print logs through pino 1. renamed logs.ts into pino.tl 2. created new plain log type
