+++ .gitignore
@@ -223,4 +223,4 @@ docs/out
 #ignore personal configs
 config.json
 #ignore databasefile
-history.db
+history.db*
