+++ src/index.ts
@@ -30,7 +30,7 @@ import boxen from 'boxen'
 import { MailData, MailgunData, SMTPData, WebhookData } from './interfaces/data'
 import { looper } from './utils/looper'
 import { printAllLogs } from './utils/logger'
-import { log } from './utils/log'
+import { log } from './utils/pino'
 import { closeLog, openLogfile, flushAllLogs } from './utils/history'
 import { notificationChecker } from './components/notification/checker'
 import { getConfig, setupConfigFromFile } from './components/config'
