+++ src/utils/pino.ts
@@ -63,6 +63,10 @@ interface NotifyLogObject extends BaseLogObject {
   url: string
 }
 
+interface PlainLogObject {
+  msg?: string
+}
+
 export type LogObject = ProbeLogObject | AlertLogObject | NotifyLogObject
 
 const isProbeLogObject = (obj: any): obj is ProbeLogObject => {
@@ -75,6 +79,10 @@ const isAlertLogObject = (obj: any): obj is AlertLogObject => {
 
 const isNotifyLogObject = (obj: any): obj is NotifyLogObject => {
   return obj?.type?.startsWith('NOTIFY')
+}
+
+const isPlainLog = (obj: any): obj is PlainLogObject => {
+  return obj
 }
 
 const prettyPrint = {
@@ -100,6 +108,9 @@ const prettyPrint = {
       return `${time},${type},${log.alertType || '-'},${
         log.notificationType || '-'
       },${log.notificationId || '-'},${log.probeId || '-'},${log.url || '-'}`
+    }
+    if (isPlainLog(log)) {
+      return `${log.msg}`
     }
     return `${time} ${log[messageKey]}`
   },
